+++ image-build/src/main/java/software/xdev/tci/imagebuild/config/BuildImageHandlerConfig.java
@@ -24,6 +24,8 @@ public interface BuildImageHandlerConfig
 {
 	boolean deleteOnExit();
 	
+	boolean addImageCacheIgnoreLabel();
+	
 	String loggerForBuildPrefix();
 	
 	Optional<String> cacheFrom();